amd-amf-vce: Fix exceptions, threading and timestamps (again).
This patch remodels the threading approach to only use a single loop, so it is much more responsive - a stall in SubmitInput or QueryOutput should no longer affect it as much while closing and debugging this should be much easier. Also possibly fixes a crash with desynchronized CPUs that somehow close the thread later than requested.
Also includes a fix for timestamps that should make files playable in VLC again, with hopefully no precision loss. Still needs additional testing.