API: Slightly rewrite API to fix 'device busy' crash
Due to an incorrect cleanup on MSVC's auto-cleanup (seems to be too far delayed), the device used would report back busy when two D3D11 or D3D9 instances are launched immediately after another in the same application. This aims at fixing that by immediately destroying the context instead of waiting for the (most likely uncontrolled) cleanup to happen.