H265: 'GOP Size' should default to FrameRate
Since 'GOP Size' is an Expert property, the plugin should always make it match the FrameRate unless it is overridden by the user. 'GOP Type' is always Fixed in firmware, so there's nothing we need to do there. Also ensures that GOP Size is not visible in anything but Expert view mode.