Encoder: Simplify & fix VBV Buffer Strictness
Constant QP could cause the VBV Buffer Strictness by default to go straight past the limit due to how it was calculated. This way should make it slightly more obvious as to how the slider works, with 50% being the halfway point for Constant QP instead of 0% and 50% being identical after going above a certain QP level.