H264: Fix 'Streaming Service Settings' override (#228)
aquamacho found a rather odd bug with the 'Enforce Streaming Service Settings' override code that would incorrectly apply an IDR Period of 0 after another one was already correctly applied.

This was because the code was testing for -1 instead of it being greater than 0, resulting in the check always running, snapping any Keyframe Interval to full second intervals, or in the reports case to 0 seconds.

This was not only restricted to Keyframe Interval, the following were also affected by this: Target Bitrate, Peak Bitrate, Rate Control Method, Profile and Quality Preset.

Fixes #228 and fixes #239.