project: Properly license under GPLv2 as required
After talking with a lawyer fluent in licensing and technology law, it turns out that MIT/X11 is not compatible with GPLv2 in a lot of cases. While this plugin linking to the AMF SDK does not break GPLv2 (AMF is a system library), the plugin itself was breaking GPLv2.