amf: Blacklist Drivers that are older than 1.4.6.0
I really didn't want to do this, as there are perfectly fine drivers. But there are some drivers that people refuse to update from which cause crashes outside of my control, which means I either have to blacklist those specific ones, or just blacklist everything that is too old.

The latter approach, blacklisting everything that is too old should work for now, and users that need to downgrade to older drivers should be linked to the AMF Plugin releases page. The versions on there go all the way back to Media SDK, but may not necessarily be bug free.