project: Update ReadMe file and remove Authors
The old ReadMe file did not convey clearly what this is and what you do with it and instead relied on people to use the wiki, while providing next to no info. The new ReadMe file includes basic instructions for installation, build status and special thanks (previously done via the Authors file).