project: Fix include groups, sorting and type
Various includes do not work correctly with clang or GCC and have to be put into an `extern "C"` section, such as windows.h. As a precaution, the obs headers were also put into an `extern "C"` section as the library is written in C.

Not that AMF is not written in C, even though it uses the `h` extension instead of `hpp`. So `extern "C"` is not used here, otherwise we would end with compile errors.