amf: Require AMF Runtime to match plugin version
The blacklist is being enabled again due to an increase in startup crashes from people that have made a intentional decision to never update their drivers, and thus will no longer be supported by new version of the plugin.

By doing this we get rid of many crashes, issues and general nonsense that just happens when outdated or broken drivers end up being used. Overall it reduces the necessary workload to get anything working.

As a result of this change, people with the stock AMD driver on Windows 8 will also now be forced to either switch to Windows 7 or Windows 10, or figure out a way to install one of those drivers on Windows 8.