cmake: Update build system for CMake build framework 3.0
This also simplifies the CMake script tremendously as it assumes the
project being built as a core obs-studio plugin.

Minimum required CMake version is 3.24, specifying `OBS_CMAKE_VERSION`
as a cache variable with a version lower than 3.0.0 switches back to the
original CMake script.

NOTE: This requires CMake build framework 3.0 on obs-studio.